backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-filter==21.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.2
Markdown==3.3.4
oauthlib==3.1.1
phonenumbers==8.12.31
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7

django-extensions==3.1.3
Werkzeug==2.0.1